Revert "Merge pull request #47 from ctidigital/develop"
This reverts commit 49f92d344f06053422021541acc5e38712f2be6a, reversing
changes made to 488cdb10d33cfbf6149fbd7a6931bb46c1f21113.
