+++ api/routes/index.js
@@ -18,8 +18,6 @@ import propertyFormImageDelete from './report/property-form-image-delete';
 import propertyFormAddress from './report/property-form-address';
 import salvageAward from './report/salvage-award';
 import checkYourAnswers from './report/check-your-answers';
-import portalDashboard from './portal/dashboard';
-import portalDashboardSort from './portal/dashboard-sort';
 
 export default () => {
   const app = Router();
@@ -43,8 +41,6 @@ export default () => {
   propertyFormAddress(app);
   salvageAward(app);
   checkYourAnswers(app);
-  portalDashboard(app);
-  portalDashboardSort(app);
 
   return app;
 };
