+++ app/js/index.js
@@ -14,5 +14,3 @@ import './components/imageUpload.js';
 import './components/spreadsheetUpload.js';
 import './components/bulkImageUpload.js';
 import './components/reportFilter.js';
-import './components/portalSignIn.js';
-import './components/portalSort.js';
