+++ app/views/report/salvage-award.html
@@ -68,8 +68,7 @@
               text: "No",
               checked: checked("claim-salvage", "no")
             }
-          ],
-          errorMessage: errors["claim-salvage"]
+          ]
         }) }}
 
         <button class="govuk-button" data-module="govuk-button">Continue</button>
