+++ app/views/report/salvaged-from.html
@@ -61,15 +61,15 @@
               text: "Afloat",
               checked: checked("removed-from", "afloat"),
               hint: {
-                text: "The material was recovered from the surface of the sea."
+                text: "The material recovered from the surface of the sea."
               }
             },
             {
-              value: "sea shore",
+              value: "ashore",
               text: "Sea shore",
-              checked: checked("removed-from", "sea shore"),
+              checked: checked("removed-from", "ashore"),
               hint: {
-                text: "The material was recovered from land such as a beach."
+                text: "The material recovered from land such as a beach."
               }
             }
           ],
