+++ package.json
@@ -25,7 +25,6 @@
     "@babel/plugin-transform-runtime": "^7.12.17",
     "@babel/preset-env": "^7.11.5",
     "accessible-autocomplete": "^2.0.3",
-    "adal-node": "^0.2.2",
     "autoprefixer": "^10.0.0",
     "axios": "^0.21.1",
     "babel-loader": "^8.1.0",
@@ -37,7 +36,6 @@
     "connect-azuretables": "^1.0.20",
     "copy-webpack-plugin": "^6.1.0",
     "css-loader": "^4.3.0",
-    "dayjs": "^1.10.4",
     "dotenv": "^8.2.0",
     "esm": "^3.2.25",
     "express": "^4.17.1",
@@ -60,7 +58,6 @@
     "postcss-loader": "^4.0.2",
     "postcss-preset-env": "^6.7.0",
     "precss": "^4.0.0",
-    "prettier": "^2.2.1",
     "sass-loader": "^10.0.2",
     "style-loader": "^1.2.1",
     "terser-webpack-plugin": "^4.2.1",
