--- utils/index.js
@@ -1,7 +1,0 @@
-export { addCheckedFunction } from './addCheckedFunction';
-export { addNunjucksFilters } from './addNunjucksFilters';
-export { forceHttps } from './forceHttps';
-export { formatValidationErrors } from './formatValidationErrors';
-export { matchRoutes } from './matchRoutes';
-export { sessionData } from './sessionData';
-export { GT_OSGB as default } from './geotools';
